% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{export2hparacne}
\alias{export2hparacne}
\title{Export hpARACNe files}
\usage{
export2hparacne(
  datl,
  output_dir,
  kinases,
  phosphatases,
  interactions = NULL,
  target_sites = NULL,
  confidence_threshold = 0,
  restrict_interactions = TRUE,
  interaction_level = "substrate"
)
}
\arguments{
\item{datl}{Unified phosphoviper format}

\item{output_dir}{Output directory}

\item{kinases}{List of kinases (UniProtKB)}

\item{phosphatases}{List of phosphatases (UniProtKB)}

\item{interactions}{HSM/D or HSM/P Interaction table (UniProtKB)}

\item{target_sites}{(Optional) list of target sites to restrict dataset (PV site_id)}

\item{confidence_threshold}{Interaction confidence_threshold}

\item{restrict_interactions}{Restrict interactions to kinase/phosphatase regulator-target subset}

\item{interaction_level}{Interaction level ("substrate" or "activity")}
}
\description{
This function exports hpARACNe input files from the unified phosphoviper format
}
