% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{export2mx}
\alias{export2mx}
\title{Export to matrix}
\usage{
export2mx(datl)
}
\arguments{
\item{datl}{Unified phosphoviper format}
}
\value{
matrix
}
\description{
This function exports tables in the unified phosphoviper format to a matrix
}
