% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{hparacne2benchmark}
\alias{hparacne2benchmark}
\title{Benchmark hpARACNe network against PathwayCommons}
\usage{
hparacne2benchmark(network, peptides, pwfile, samplefactor = 1)
}
\arguments{
\item{network}{hpARACNe network}

\item{pwfile}{PathwayCommons SIF file}

\item{negfile}{Negative interactome file}
}
\description{
This function conducts a benchmark of a hpARACNe network against PathwayCommons
}
