% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{hparacne2regulon}
\alias{hparacne2regulon}
\title{Import hpARACNe network}
\usage{
hparacne2regulon(
  afile,
  pfile,
  mfile = NA,
  method = "spearman",
  likelihood_threshold = 0.5,
  priors = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{afile}{hpARACNe network file}

\item{pfile}{hpARACNe peptides file}

\item{mfile}{Optional hpARACNe quantitative matrix file for TFmode refinement}

\item{method}{Correlation method for TFmode refinement}

\item{likelihood_threshold}{Interaction confidence threshold to filter likelihood}

\item{priors}{Logical, whether prior interaction probabilities should be used as weights for likelihood}

\item{verbose}{Logical, whether progression messages should be printed in the terminal}
}
\description{
This function imports a hpARACNe network and generates site-specific regulons
}
