% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importCCT}
\alias{importCCT}
\title{Import CCT file}
\usage{
importCCT(file)
}
\arguments{
\item{file}{CPTAC file}
}
\value{
peptide-level phosphoviper data.table
}
\description{
This function imports a CCT file and converts the data to the unified phosphoviper format
}
