% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importCPTAC}
\alias{importCPTAC}
\title{Import CPTAC file}
\usage{
importCPTAC(file, fasta, cores = 1)
}
\arguments{
\item{file}{CPTAC file}

\item{fasta}{Amino acid FASTA file from UniProt}

\item{cores}{Integer indicating the number of cores to use (only 1 in Windows-based systems)}
}
\value{
peptide-level phosphoviper data.table
}
\description{
This function imports a CPTAC file and converts the data to the unified phosphoviper format
}
