% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importPhosphoCCT}
\alias{importPhosphoCCT}
\title{Import phospho CCT file}
\usage{
importPhosphoCCT(file)
}
\arguments{
\item{file}{CCT file}
}
\value{
peptide-level phosphoviper data.table
}
\description{
This function imports a phospho CCT file and converts the data to the unified phosphoviper format
}
