% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importPhosphoSANGER}
\alias{importPhosphoSANGER}
\title{Import phospho SANGER file}
\usage{
importPhosphoSANGER(file, fasta, cores = 1)
}
\arguments{
\item{file}{phospho SANGER file}

\item{fasta}{Amino acid FASTA file from UniProt}

\item{cores}{Integer indicating the number of cores to use (only 1 in Windows-based systems)}
}
\value{
peptide-level vespa data.table
}
\description{
This function imports a phospho SANGER file and converts the data to the unified vespa format
}
