% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importProteoCCT}
\alias{importProteoCCT}
\title{Import protein CCT file}
\usage{
importProteoCCT(file, fasta, hgnc)
}
\arguments{
\item{file}{protein-level CCT file}

\item{fasta}{Amino acid FASTA file from UniProt}

\item{hgnc}{HGNC mapping file}
}
\value{
peptide-level phosphoviper data.table
}
\description{
This function imports a protein CCT file and converts the data to the unified phosphoviper format
}
