% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importProteoMaxQuant}
\alias{importProteoMaxQuant}
\title{Import protein MaxQuant proteinGroups TXT file}
\usage{
importProteoMaxQuant(file, fasta, cores = 1)
}
\arguments{
\item{file}{protein MaxQuant proteinGroups TXT file}

\item{fasta}{Amino acid FASTA file from UniProt}

\item{cores}{Integer indicating the number of cores to use (only 1 in Windows-based systems)}
}
\value{
peptide-level phosphoviper data.table
}
\description{
This function imports a protein MaxQuant proteinGroups TXT file and converts the data to the unified phosphoviper format
}
