% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importProteoProgenesis}
\alias{importProteoProgenesis}
\title{Import protein Progenesis TXT file}
\usage{
importProteoProgenesis(file, run_ids)
}
\arguments{
\item{file}{proteo Progenesis TXT file}

\item{run_ids}{Column names of runs to extract}
}
\value{
peptide-level phosphoviper data.table
}
\description{
This function imports a protein Progenesis TXT file and converts the data to the unified phosphoviper format
}
