% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importProteoSANGER}
\alias{importProteoSANGER}
\title{Import protein SANGER file}
\usage{
importProteoSANGER(file)
}
\arguments{
\item{file}{protein SANGER file}
}
\value{
peptide-level phosphoviper data.table
}
\description{
This function imports a protein SANGER file and converts the data to the unified phosphoviper format
}
