% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importSANGER}
\alias{importSANGER}
\title{Import SANGER file}
\usage{
importSANGER(file, fasta, cores = 1)
}
\arguments{
\item{file}{SANGER file}

\item{fasta}{Amino acid FASTA file from UniProt}

\item{cores}{Integer indicating the number of cores to use (only 1 in Windows-based systems)}
}
\value{
peptide-level phosphoviper data.table
}
\description{
This function imports a SANGER file and converts the data to the unified phosphoviper format
}
