% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importSpectronaut}
\alias{importSpectronaut}
\title{Import Spectronaut TXT file}
\usage{
importSpectronaut(file, fasta, run_ids, cores = 1)
}
\arguments{
\item{file}{Spectronaut TXT file}

\item{fasta}{Amino acid FASTA file from UniProt}

\item{run_ids}{Column names of runs to extract}

\item{cores}{Integer indicating the number of cores to use (only 1 in Windows-based systems)}
}
\value{
peptide-level phosphoviper data.table
}
\description{
This function imports a Spectronaut TXT file and converts the data to the unified phosphoviper format
}
