% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper.R
\name{iterviper}
\alias{iterviper}
\title{Run VIPER iteratively}
\usage{
iterviper(eset, regulon, iterations, ...)
}
\arguments{
\item{datl}{Unified phosphoviper format}
}
\value{
matrix
}
\description{
This function runs VIPER iteratively
}
