% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{loadHSMPF}
\alias{loadHSMPF}
\title{Load human reference HSM/P interactions}
\usage{
loadHSMPF()
}
\value{
Table of PPI with their scores
}
\description{
Based on Cunningham JM, Koytiger G, Sorger PK, AlQuraishi M. Nat Methods. 2020 Feb;17(2):175-183. doi: 10.1038/s41592-019-0687-1.
}
\references{
\url{https://figshare.com/articles/Predictions_-_Domain-Peptide_and_Protein-Protein_Interactions_-_HSM/10084745/2}
}
