% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{loadHumanKinases}
\alias{loadHumanKinases}
\title{Load human reference kinases}
\usage{
loadHumanKinases()
}
\value{
List of human kinases in UniProtKB format
}
\description{
Based on pkinfam version 2018_03
}
\references{
\url{https://www.uniprot.org/docs/pkinfam}
}
