% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{loadHumanPhosphatases}
\alias{loadHumanPhosphatases}
\title{Load human reference phosphatases}
\usage{
loadHumanPhosphatases()
}
\value{
List of human phosphatases in UniProtKB format
}
\description{
Based on Li X, Wilmanns M, Thornton J, Köhn M. Sci Signal. 2013 May 14;6(275):rs10. doi: 10.1126/scisignal.2003203.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pubmed/23674824}
}
