% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{loadPCdb}
\alias{loadPCdb}
\title{Load human reference PathwayCommons interactions}
\usage{
loadPCdb()
}
\value{
Table of PPI with their scores
}
\description{
Based on PathwayCommons, version 12.
}
\references{
\url{https://www.pathwaycommons.org/}
}
