% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{optimizeRegulon}
\alias{optimizeRegulon}
\title{metaVIPER regulon selection}
\usage{
optimizeRegulon(ges, net.list, min_size = 10)
}
\arguments{
\item{ges}{Gene Expression Signature (features X samples)}

\item{net.list}{List object with the networks to be used}
}
\value{
Optimized network with best regulon per gene
}
\description{
metaVIPER regulon selection
}
