% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{orthogonalRegulon}
\alias{orthogonalRegulon}
\title{Orthogonal site-specific regulon selection}
\usage{
orthogonalRegulon(ges, regulon, min_size = 10, cutoff = 0.5)
}
\arguments{
\item{ges}{Gene Expression Signature (features X samples)}

\item{regulon}{Site-specific regulons}

\item{min_size}{Minimum regulon size}

\item{cutoff}{Pairwise correlation cutoff to remove regulons}
}
\value{
Optimized network containing only orthogonal site-specific regulons for each protein
}
\description{
Orthogonal site-specific regulon selection
}
