% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphoviper.R
\docType{package}
\name{phosphoviper}
\alias{phosphoviper}
\title{phosphoVIPER: phosphoproteomic extension for VIPER}
\description{
phosphoVIPER provides extensions to analyse phosphoproteomic data using hpARACNe and VIPER.
}
\details{
\tabular{ll}{
Package: \tab phosphoviper\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
License: \tab Columbia University Non-Commercial License\cr
}
}
\author{
George Rosenberger \email{gr2578@cumc.columbia.edu}
}
