% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_fairing.R
\name{region_fairing}
\alias{region_fairing}
\title{Fair a region of a triangular surface mesh}
\usage{
region_fairing(mesh, point_ids, update_attributes = TRUE)
}
\arguments{
\item{mesh}{A \code{mesh3d} object.}

\item{point_ids}{Integer vector of 1-based vertex indices defining the region
to be faired.}

\item{update_attributes}{Whether to interpolate point and cell data from the
input. Default \code{TRUE}.}
}
\value{
A \code{mesh3d} object with the selected region faired.
}
\description{
Smooths a user-defined region of a triangulated surface mesh using CGAL's
region fairing algorithm. The region vertices are moved to minimise a
local fairing energy while the boundary of the region is kept fixed.
}
\examples{
f <- system.file("extdata", "torus.stl", package = "vespa")
mesh <- read_stl(f)
result <- region_fairing(mesh, point_ids = 1:5)

}
\seealso{
Other mesh_transforms:
\code{\link[=alpha_wrapping]{alpha_wrapping()}},
\code{\link[=boolean_operation]{boolean_operation()}},
\code{\link[=isotropic_remeshing]{isotropic_remeshing()}},
\code{\link[=mesh_check]{mesh_check()}},
\code{\link[=mesh_deformation]{mesh_deformation()}},
\code{\link[=mesh_smoothing]{mesh_smoothing()}},
\code{\link[=mesh_subdivision]{mesh_subdivision()}},
\code{\link[=patch_filling]{patch_filling()}},
\code{\link[=pca_estimate_normals]{pca_estimate_normals()}},
\code{\link[=poisson_reconstruction]{poisson_reconstruction()}},
\code{\link[=shape_smoothing]{shape_smoothing()}}
}
\concept{mesh_transforms}
