% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{regulator2protein}
\alias{regulator2protein}
\title{Convert site-specific regulators to protein regulators}
\usage{
regulator2protein(regulons)
}
\arguments{
\item{regulons}{VIPER regulons}
}
\description{
This function converts site-specific regulators to site-specific regulators
}
