% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{regulons2benchmark}
\alias{regulons2benchmark}
\title{Benchmark regulons against PathwayCommons}
\usage{
regulons2benchmark(regulons, pwfile)
}
\arguments{
\item{regulons}{Gene-level VIPER regulons}

\item{pwfile}{PathwayCommons SIF file}
}
\description{
This function conducts a benchmark of gene regulons against PathwayCommons
}
