% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{site2gene}
\alias{site2gene}
\title{Convert site-specific regulons to gene regulons}
\usage{
site2gene(regulons)
}
\arguments{
\item{regulons}{VIPER regulons}
}
\description{
This function converts site-specific regulons to a list of gene - gene regulons
}
