% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{subsetRegulon}
\alias{subsetRegulon}
\title{Subset regulons for targets}
\usage{
subsetRegulon(regulons, targets, min_size = 10)
}
\arguments{
\item{regulons}{VIPER regulons}

\item{min_size}{minimum regulon size}

\item{target}{Vector of targets}
}
\description{
This function subsets regulons to specified targets only
}
