% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{summary.rocl}
\alias{summary.rocl}
\title{Benchmark regulons against PathwayCommons}
\usage{
\method{summary}{rocl}(rocl)
}
\arguments{
\item{rocl}{Benchmark object}
}
\description{
Print summary statistics for benchmark
}
