% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testQuadrat.R
\name{testQuadrat}
\alias{testQuadrat}
\title{quadrat test}
\arguments{
\item{x}{\code{data.frame} or \code{\link{SingleCellExperiment}}.}

\item{xy}{character vector;
specifies the variables containing 
spatial (x- and y-)coordinates.}

\item{BPPARAM}{a \code{\link[BiocParallel]{BiocParallelParam}}
instance defining the parallel back-end to be used.}

\item{by}{character vector; 
specifies a variable or set of variables to group points by. 
Each grouping will be tested separately.}

\item{...}{further arguments;
passed to \code{\link[spatstat.core]{quadrat.test}}.}
}
\value{
a \code{\link[S4Vectors]{DataFrame-class}} object.
}
\description{
Test for complete spatial randomness (CSR) of points based on 
quadrat counts (see \code{\link[spatstat.core]{quadrat.test}}).
}
\details{
In brief, the window of observation is divided into \eqn{n} tiles, 
and the observed number of data points \eqn{O} in each tile \eqn{i} 
is counted. Then, the expected number of points in each quadrat \eqn{E} 
is calculated, and the Pearson statistic computed according to: 
\deqn{\chi^2=\sum_{i=1}^n((O_i-E)^2/E)}
}
\examples{
data(foo)
testQuadrat(foo, by = "subset")

}
\author{
Helena L. Crowell
}
