% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vespa.R
\docType{package}
\name{vespa}
\alias{vespa}
\title{vespa: Virtual Enrichment-based Signaling Protein-activity Analysis}
\description{
VESPA provides extensions to analyse phosphoproteomic data by ARACNe and VIPER.
}
\details{
\tabular{ll}{
Package: \tab vespa\cr
Type: \tab Package\cr
License: \tab Columbia University Non-Commercial License\cr
}
}
\author{
George Rosenberger \email{gr2578@cumc.columbia.edu}
}
