% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_fairing.R
\name{vespa_region_fair}
\alias{vespa_region_fair}
\title{Fair a region of a triangular surface mesh}
\usage{
vespa_region_fair(mesh, point_ids, update_attributes = TRUE)
}
\arguments{
\item{mesh}{A \code{mesh3d} object.}

\item{point_ids}{Integer vector of 1-based vertex indices defining the region
to be faired.}

\item{update_attributes}{Whether to interpolate point and cell data from the
input. Default \code{TRUE}.}
}
\value{
A \code{mesh3d} object with the selected region faired.
}
\description{
Smooths a user-defined region of a triangulated surface mesh using CGAL's
region fairing algorithm. The region vertices are moved to minimise a
local fairing energy while the boundary of the region is kept fixed.
}
