% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{vmx2pv}
\alias{vmx2pv}
\title{Export to unified vespa format from viper matrix}
\usage{
vmx2pv(vmx, fasta = NULL, tag = "PV")
}
\arguments{
\item{fasta}{Amino acid FASTA file from UniProt}

\item{tag}{Site tag}

\item{matrix}{phosphoVIPER quantitative matrix}
}
\value{
vespa table
}
\description{
This function exports a viper matrix to the unified vespa format
}
