% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{export2hparacne}
\alias{export2hparacne}
\title{Export hpARACNe files}
\usage{
export2hparacne(
  datl,
  output_dir,
  kinases,
  phosphatases,
  interactions,
  confidence_threshold = 0.5
)
}
\arguments{
\item{datl}{Unified phosphoviper format}

\item{output_dir}{Output directory}

\item{kinases}{List of kinases (UniProtKB)}

\item{phosphatases}{List of phosphatases (UniProtKB)}

\item{interactions}{HSM/D or HSM/P Interaction table (UniProtKB)}

\item{confidence_threshold}{Interaction confidence_threshold}
}
\description{
This function exports hpARACNe input files from the unified phosphoviper format
}
