% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{export2hparacne}
\alias{export2hparacne}
\title{Export hpARACNe files}
\usage{
export2hparacne(datl, output_dir, kinases, phosphatases)
}
\arguments{
\item{datl}{Unified phosphoviper format}

\item{output_dir}{Output directory}

\item{kinases}{List of kinases (UniProtKB)}

\item{phosphatases}{List of phosphatases (UniProtKB)}
}
\description{
This function exports hpARACNe input files from the unified phosphoviper format
}
