% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{export2mx}
\alias{export2mx}
\title{Export to matrix}
\usage{
export2mx(datl, fillvalues = NA)
}
\arguments{
\item{datl}{Unified phosphoviper format}

\item{fillvalues}{Missing value imputation method ("rowmin": Row-wise (peptide) minimum; "colmin": Column-wise (run) minimum; NULL: Skip and use NA)}
}
\value{
matrix
}
\description{
This function exports tables in the unified phosphoviper format to a matrix
}
