% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{export2mx}
\alias{export2mx}
\title{Export to matrix}
\usage{
export2mx(datl, fillvalues = NA, jitter = TRUE)
}
\arguments{
\item{datl}{Unified vespa format}

\item{fillvalues}{Missing value imputation method ("rowmin": Row-wise (peptide) minimum; "colmin": Column-wise (run) minimum; NULL: Skip and use NA)}

\item{jitter}{Add numerical noise to imputed missing values}
}
\value{
matrix
}
\description{
This function exports tables in the unified vespa format to a matrix
}
