% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_isosurface.R
\name{extract_isosurface}
\alias{extract_isosurface}
\alias{extract_isosurface.sdf_volume}
\alias{extract_isosurface.default}
\title{Extract an isosurface from a 3-D scalar field}
\usage{
extract_isosurface(x, isovalue = 0, ...)

\method{extract_isosurface}{sdf_volume}(x, isovalue = 0, ...)

\method{extract_isosurface}{default}(x, isovalue = 0, ...)
}
\arguments{
\item{x}{An object from which to extract an isosurface.}

\item{isovalue}{Numeric. The iso-level at which to extract the surface.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{mesh3d} object (from package \strong{rgl}).
}
\description{
Generic function that extracts an isosurface from a volumetric scalar
field. Currently implemented for objects of class \code{\link{sdf_volume}}
(returned by \code{\link[=signed_distance_function]{signed_distance_function()}}).
}
\seealso{
\code{\link[=signed_distance_function]{signed_distance_function()}} to build an \code{sdf_volume}.
}
