% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{hparacne2regulon}
\alias{hparacne2regulon}
\title{Import hpARACNe network}
\usage{
hparacne2regulon(afile, pfile)
}
\arguments{
\item{afile}{hpARACNe network file}

\item{pfile}{hpARACNe peptides file}
}
\description{
This function imports a hpARACNe network and generates site-specific regulons
}
