% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{hparacne2regulon}
\alias{hparacne2regulon}
\title{Import hpARACNe network}
\usage{
hparacne2regulon(afile, pfile, verbose = TRUE)
}
\arguments{
\item{afile}{hpARACNe network file}

\item{pfile}{hpARACNe peptides file}

\item{verbose}{Logical, whether progression messages should be printed in the terminal}
}
\description{
This function imports a hpARACNe network and generates site-specific regulons
}
