% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importOpenSWATH}
\alias{importOpenSWATH}
\title{Import OpenSWATH TSV file}
\usage{
importOpenSWATH(
  file,
  fasta,
  normalization_method = FALSE,
  normalization_center = TRUE,
  batchfile,
  cores = 1
)
}
\arguments{
\item{file}{OpenSWATH TSV file}

\item{fasta}{Amino acid FASTA file from UniProt}

\item{normalization_method}{Either "FALSE" (skip normalization), "quantile" or "cyclicLoess" normalization}

\item{normalization_center}{Either FALSE (skip center normalization) or TRUE}

\item{batchfile}{A data.table with tags (replacement for run_id) and batch annotation for separate normalization (columns: run_id, tag, aggregator_id, ds_id)}

\item{cores}{Integer indicating the number of cores to use}
}
\value{
peptide-level phosphoviper data.table
}
\description{
This function imports a OpenSWATH TSV file and converts the data to the unified phosphoviper format
}
