% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importOpenSWATH}
\alias{importOpenSWATH}
\title{Import OpenSWATH TSV file}
\usage{
importOpenSWATH(
  file,
  fasta,
  batchfile,
  cores = 1,
  normalization_method = FALSE,
  batchcorrection = FALSE,
  batchsubset = FALSE,
  batchlevel = c("protein_id", "peptide_id"),
  min_detections = 10
)
}
\arguments{
\item{file}{OpenSWATH TSV file}

\item{fasta}{Amino acid FASTA file from UniProt}

\item{batchfile}{A data.table with batch correction annotation (columns: run_id, tag, aggregator_id, ds_id)}

\item{cores}{Integer indicating the number of cores to use (only 1 in Windows-based systems)}

\item{normalization_method}{Either "FALSE" (skip normalization), "quantile" or "cyclicLoess" normalization}

\item{batchcorrection}{Whether batch subsetting should be conducted}

\item{batchlevel}{Minimum number of peptide detections across samples}
}
\value{
peptide-level phosphoviper data.table
}
\description{
This function imports a OpenSWATH TSV file and converts the data to the unified phosphoviper format
}
