% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importProteoCPTAC}
\alias{importProteoCPTAC}
\title{Import protein CPTAC file}
\usage{
importProteoCPTAC(file, fasta)
}
\arguments{
\item{file}{protein-level CPTAC file}

\item{fasta}{Amino acid FASTA file from UniProt}
}
\value{
peptide-level vespa data.table
}
\description{
This function imports a protein CPTAC file and converts the data to the unified vespa format
}
