% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{optimizeRegulon}
\alias{optimizeRegulon}
\title{metaVIPER regulon selection}
\usage{
optimizeRegulon(
  ges,
  net.list,
  min_size = 25,
  pleiotropy = FALSE,
  pleiotropyArgs = list(regulators = 0.05, shadow = 0.05, targets = 10, penalty = 20,
    method = "adaptive")
)
}
\arguments{
\item{ges}{Gene Expression Signature (features X samples)}

\item{net.list}{List object with the networks to be used}
}
\value{
Optimized network with best regulon per gene
}
\description{
metaVIPER regulon selection
}
