% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{site2protein}
\alias{site2protein}
\title{Convert site-specific regulons to protein regulons}
\usage{
site2protein(regulons, average_regulons = TRUE)
}
\arguments{
\item{regulons}{VIPER regulons}
}
\description{
This function converts site-specific regulons to a list of protein - protein regulons
}
