% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{subset_regulon}
\alias{subset_regulon}
\title{Subset regulons for targets}
\usage{
subset_regulon(regulons, targets)
}
\arguments{
\item{regulons}{VIPER regulons}

\item{target}{Vector of targets}
}
\description{
This function subsets regulons to specified targets only
}
