% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testColo.R
\name{testColo}
\alias{testColo}
\title{co-localization}
\arguments{
\item{x}{\code{data.frame} or \code{\link{SingleCellExperiment}}.}

\item{xy}{character vector;
specifies the variables containing 
spatial (x- and y-)coordinates.}

\item{BPPARAM}{a \code{\link[BiocParallel]{BiocParallelParam}}
instance defining the parallel back-end to be used.}

\item{...}{further arguments; passed to \code{\link[ks]{kde}}.}

\item{by}{character string; 
specifies the variable to test by for co-localization.}

\item{group}{character vector; 
optional additional variable(s) to group by, e.g., 
in order to perform a series of independent tests.}

\item{nc}{scalar integer; 
threshold on the number of cells above 
which to consider a group for testing.}
}
\value{
a \code{\link[S4Vectors]{DataFrame-class}} object.
}
\description{
...
}
\details{
...
}
\examples{
data(foo)
testColo(foo, 
  by = "subset", 
  gridsize = c(5, 5))

}
\author{
Helena L. Crowell
}
