% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vespa-package.R
\docType{package}
\name{vespa-package}
\alias{vespa}
\alias{vespa-package}
\title{vespa: R Bindings for the Vespa VTK/CGAL Mesh Processing Library}
\description{
Provides R wrappers for all Vespa filters (isotropic remeshing, boolean operations, alpha wrapping, surface reconstruction, point set processing, and more) via direct Rcpp bindings to the Vespa C++/VTK library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cregouby.github.io/vespa/}
}

}
\author{
\strong{Maintainer}: Christophe Regouby \email{christophe.regouby@airbus.com}

Authors:
\itemize{
  \item Christophe Regouby \email{christophe.regouby@airbus.com}
}

}
\keyword{internal}
